[Fix] post_save not working
